core: livequery's tests move to the child module file
The maintainer's convention puts a module's test module in a sibling
tests.rs, as core/src/event_dag already does: the parent keeps
`#[cfg(test)] mod tests;` and livequery/ holds the body. Pure relocation
of the module the previous commit introduced - no test text changed
beyond the four spaces of nesting each line loses.
